bulk WriteDistance, FTLSupport nearest node javítás
git-tfs-id: [https://gyorgyiantal.visualstudio.com]$/PMap/PMap;C618

+++ PMap/BO/boRoute.cs
@@ -23,7 +23,7 @@ namespace PMap.BO
         public int DST_MAXWEIGHT { get; set; }
         public int DST_MAXHEIGHT { get; set; }
         public int DST_MAXWIDTH { get; set; }
-        public double DST_DISTANCE { get; set; }
+        public int DST_DISTANCE { get; set; }
         public double CalcDistance                  //lehet, hogy nem hasznaljuk
         {
             get
